+++ syndicate/core/handlers.py
@@ -78,7 +78,7 @@ def syndicate():
 @click.option('--project_path', type=str,
               help='Path to project folder. Default value: working dir')
 # account settings
-@click.option('--account_id', callback=check_required_param, type=int,
+@click.option('--account_id', callback=check_required_param, type=str,
               help='[required] Id of the AWS account where to deploy '
                    'application')
 @click.option('--region', type=str, default='us-west-1',
@@ -310,12 +310,13 @@ def update(bundle_name, deploy_name, replace_output,
 @click.option('--clean_only_types', multiple=True)
 @click.option('--clean_only_resources', multiple=True)
 @click.option('--clean_only_resources_path', nargs=1, type=str)
+@click.option('--clean_externals', nargs=1, is_flag=True, default=False)
 @click.option('--excluded_resources', multiple=True)
 @click.option('--excluded_resources_path', nargs=1, type=str)
 @click.option('--excluded_types', multiple=True)
 @click.option('--rollback', is_flag=True)
 def clean(deploy_name, bundle_name, clean_only_types, clean_only_resources,
-          clean_only_resources_path, excluded_resources,
+          clean_only_resources_path, clean_externals, excluded_resources,
           excluded_resources_path, excluded_types, rollback):
     """
     Cleans the application infrastructure.
@@ -326,6 +327,7 @@ def clean(deploy_name, bundle_name, clean_only_types, clean_only_resources,
         must be cleaned
     :param clean_only_resources_path: path to a json list which contains a list
         of resources names which must be cleaned
+    :param clean_externals: used to clean external resources
     :param excluded_resources: list of the resources names than must be skipped
         while cleaning the application infrastructure
     :param excluded_resources_path: path to a json list which contains a list
@@ -367,14 +369,16 @@ def clean(deploy_name, bundle_name, clean_only_types, clean_only_resources,
                                        clean_only_resources=clean_only_resources,
                                        clean_only_types=clean_only_types,
                                        excluded_resources=excluded_resources,
-                                       excluded_types=excluded_types)
+                                       excluded_types=excluded_types,
+                                       clean_externals=clean_externals)
     else:
         remove_deployment_resources(deploy_name=deploy_name,
                                     bundle_name=bundle_name,
                                     clean_only_resources=clean_only_resources,
                                     clean_only_types=clean_only_types,
                                     excluded_resources=excluded_resources,
-                                    excluded_types=excluded_types)
+                                    excluded_types=excluded_types,
+                                    clean_externals=clean_externals)
     click.echo('AWS resources were removed.')